+++ lib/razor/task.rb
@@ -6,22 +6,22 @@ module Razor
   class TaskInvalidError < RuntimeError; end
 
   # An task is a collection of templates, plus some metadata. The
-  # metadata lives in a YAML file, the templates in a subdirectory with the
-  # same base name as the YAML file. For an task with name +name+, the
-  # YAML data must be in +name.yaml+ somewhere on
+  # task lives in a folder called +<task_name>.task+ The
+  # metadata lives inside this folder in a YAML file called +metadata.yaml+.
+  # The templates live in the task directory. For example, a task with name
+  # +name+ must have its YAML metadata in +name.task/metadata.yaml+ somewhere on
   # +Razor.config["task_path"]+.
   #
   # Templates are looked up from the directories listed in
   # +Razor.config["task_path"]+, first in a subdirectory
-  # +name/os_version+, then +name+ and finally in the fixed directory
-  # +common+.
+  # +name.task+, then in the fixed directory +common+.
   #
   # The following entries from the YAML file are used:
   # +os_version+: the OS version this task supports
   # +label+, +description+: human-readable information
   # +boot_sequence+: a hash mapping integers or the string +"default"+ to
   # template names. When booting a node, the task will respond with
-  # the numered entries in +boot_sequence+ in order; if the number of boots
+  # the numbered entries in +boot_sequence+ in order; if the number of boots
   # that a node has done under this task is not in +boot_sequence+,
   # the template marked as +"default"+ will be used
   #
@@ -58,8 +58,7 @@ module Razor
     end
 
     def find_file(filename)
-      candidates = [File::join(name, os_version, filename), File::join(name, filename)]
-      self.class.find_on_task_paths(*candidates) or
+      self.class.find_on_task_paths(Pathname(name + '.task') + filename) or
         (@base and @base.find_file(filename)) or
         self.class.find_common_file(filename) or
         raise TemplateNotFoundError, _("Task %{name}: %{filename} not on the search path") % {name: name, filename: filename}
@@ -80,18 +79,26 @@ module Razor
 
     protected :metadata
 
-    # Look up an task by name. We support file-based tasks
+    # Look up a task by name. We support file-based tasks
     # (mostly for development) and tasks stored in the database. If
     # there is both a file-based and a DB-backed task with the same
     # name, we use the file-based one.
     def self.find(name)
-      if yaml = find_on_task_paths("#{name}.yaml")
-        metadata = YAML::load(File::read(yaml)) || {}
+      if metadata_file = find_on_task_paths(Pathname(name + '.task') + 'metadata.yaml')
+        metadata = YAML::load(File.read(metadata_file)) || {}
         new(name, metadata)
       elsif inst = Razor::Data::Task[:name => name]
         inst
-      elsif !yaml
-        raise TaskNotFoundError, _("No task %{name}.yaml on search path") % {name: name}
+      elsif is_legacy_format(name)
+        # Migration error.
+        raise raise TaskNotFoundError, _(<<-EOT) % {name: name}
+Task %{name} appears to be in a legacy format:
+We found a '%{name}.yaml' file, but modern tasks are in folders named '%{name}.task'
+containing file 'metadata.yaml'.  For more detail see migration documentation online at
+http://links.puppetlabs.com/razor-migration-task-revamp
+EOT
+      else
+        raise TaskNotFoundError, _("Could not find task %{name} on the search path") % {name: name}
       end
     end
 
@@ -104,16 +111,20 @@ module Razor
     end
 
     def self.find_common_file(filename)
-      find_on_task_paths(File::join("common", filename))
+      find_on_task_paths(Pathname('common') + filename)
     end
 
     # List all known, valid tasks, both from the DB and the file
     # system. Return an array of +Razor::Task+ objects, sorted by
     # +name+
     def self.all
-      (Razor.config.task_paths.map do |ip|
-        Dir.glob(File::join(ip, "*.yaml")).map { |p| File::basename(p, ".yaml") }
-       end + Razor::Data::Task.all.map { |t| t.name }).flatten.uniq.sort.map do |name|
+      (Razor.config.task_paths.map do |search_path|
+        search_pathname = Pathname(search_path)
+        Pathname.glob(Pathname(search_path) + '**' + '*.task').map do |task_folder_pathname|
+          # Remove the base path and .task extension to get task name.
+          task_folder_pathname.relative_path_from(search_pathname).sub_ext('').to_s
+        end
+      end + Razor::Data::Task.all.map(&:name)).flatten.sort.uniq.map do |name|
         begin
           find(name)
         rescue TaskInvalidError
@@ -123,14 +134,17 @@ module Razor
     end
 
     private
-    def self.find_on_task_paths(*paths)
-      Razor.config.task_paths.each do |ip|
-        paths.each do |path|
-          fname = File::join(ip, path)
-          return fname if File::exists?(fname)
-        end
+    def self.find_on_task_paths(file)
+      Razor.config.task_paths.each do |task_path|
+        fname = File::join(task_path, file)
+        return fname if File::exists?(fname)
       end
       nil
+    end
+
+    private
+    def self.is_legacy_format(name)
+      return !find_on_task_paths(name + '.yaml').nil?
     end
   end
 end