+++ spec/data/task_spec.rb
@@ -55,6 +55,27 @@ describe Razor::Data::Task do
       }.to raise_error(Sequel::ValidationFailed)
     end
 
+    describe "task" do
+      it "should allow '/' in the middle of a name" do
+        task.set(name: 't/1').save
+      end
+      it "should reject '/' at the end of a name" do
+        rejects 'name', 't/'
+      end
+      it "should reject '/' at the start of a name" do
+        rejects 'name', '/1'
+      end
+      it "should reject '$' in the middle of a name" do
+        rejects 'name', 't$1'
+      end
+      it "should reject '$' at end of a name" do
+        rejects 'name', 't$'
+      end
+      it "should reject '$' at start of a name" do
+        rejects 'name', '$1'
+      end
+    end
+
     describe "templates" do
       def rejects_templates(value)
         rejects "templates", value